Update transformations.
This also carries out an optimization where we do the viewport as (-1,1) in x and y, but apply a transformation to squish the screen at the top of our calculations. We also alter the near and far clipping plane values to handle our perspective calculations, and we create an orthographic projection to flatten in z.